requirements.txt
asgiref
backports.zoneinfo;python_version<"3.9"
Django
django-ckeditor
django-crispy-forms
django-js-asset
Pillow
sqlparse
tzdata